+++ package.json
@@ -31,6 +31,8 @@
   },
   "dependencies": {
     "apollo-server": "^2.1.0",
-    "graphql": "^14.0.2"
+    "graphql": "^14.0.2",
+    "graphql-yoga": "^1.16.7",
+    "prisma-client-lib": "^1.21.0"
   }
 }
