+++ prisma.yml
@@ -1,0 +1,5 @@
+endpoint: http://localhost:4466
+datamodel: datamodel.prisma
+generate:
+  - generator: javascript-client
+    output: ./generated/prisma-client/
